updated gps and editing and removal of table rows
A loading indicator will now pop up when searching for gps coordinates.
If accurate enough coordinates are not found an alert dialog will notify
the user. The user can now edit through long click and a delete button
is in the dialog if they with to delete. also fixed gps so that it stops
checking once gps coordinates have been found or a timeout occurs.